Move tests and fix existing test
