+++ qa/src/test/java/io/camunda/migrator/qa/SkipAndRetryProcessInstancesTest.java
@@ -38,9 +38,6 @@ import org.springframework.test.context.TestPropertySource;
 @TestPropertySource(properties = { "logging.level.io.camunda.migrator=WARN" })
 class SkipAndRetryProcessInstancesTest extends RuntimeMigrationAbstractTest {
 
-  @RegisterExtension
-  protected final LogCapturer logs = LogCapturer.create().captureForType(RuntimeMigrator.class);
-
   @Autowired
   private RuntimeService runtimeService;
 
@@ -87,47 +84,6 @@ class SkipAndRetryProcessInstancesTest extends RuntimeMigrationAbstractTest {
     List<IdKeyDbModel> skippedProcessInstanceIds = idKeyMapper.findSkipped().stream().toList();
     assertThat(skippedProcessInstanceIds.size()).isEqualTo(1);
     assertThat(skippedProcessInstanceIds.getFirst().id()).isEqualTo(process.getId());
-  }
-
-  @Test
-  public void shouldSkipOnMissingC8Deployment() {
-    // given
-    deployCamunda7Process("io/camunda/migrator/bpmn/c7/simpleProcess.bpmn");
-    var c7Instance = runtimeService.startProcessInstanceByKey("simpleProcess");
-
-    // when
-    runtimeMigrator.start();
-
-    // then
-    logs.assertContains(String.format(
-        "Process instance with legacyId [%s] can't be migrated: "
-            + "No C8 deployment found for process ID [%s] required for instance with "
-            + "legacyID [%s].",  c7Instance.getId(), "simpleProcess", c7Instance.getId()));
-    assertThat(camundaClient.newProcessInstanceSearchRequest().send().join().items().size()).isEqualTo(0);
-    List<IdKeyDbModel> skippedProcessInstanceIds = idKeyMapper.findSkipped().stream().toList();
-    assertThat(skippedProcessInstanceIds.size()).isEqualTo(1);
-    assertThat(skippedProcessInstanceIds.getFirst().id()).isEqualTo(c7Instance.getId());
-  }
-
-  @Test
-  public void shouldSkipOnMissingElementInC8Deployment() {
-    // given an instance that is currently in an element in the C7 model which does not exist in the C8 model
-    deployProcessInC7AndC8("userTaskProcessWithMissingUserTaskInC8.bpmn");
-    var c7Instance = runtimeService.startProcessInstanceByKey("userTaskProcessWithMissingUserTaskInC8Id");
-
-    // when
-    runtimeMigrator.start();
-
-    // then
-    logs.assertContains(String.format(
-        "Process instance with legacyId [%s] can't be migrated: " +
-        "C7 instance detected which is currently in a C7 model element which does not exist in the equivalent deployed C8 model. "
-            + "Instance legacyId: [%s], Model legacyId: [%s], Element Id: [%s].",
-        c7Instance.getId(), c7Instance.getId(), "userTaskProcessWithMissingUserTaskInC8Id", "userTaskId"));
-    assertThat(camundaClient.newProcessInstanceSearchRequest().send().join().items().size()).isEqualTo(0);
-    List<IdKeyDbModel> skippedProcessInstanceIds = idKeyMapper.findSkipped().stream().toList();
-    assertThat(skippedProcessInstanceIds.size()).isEqualTo(1);
-    assertThat(skippedProcessInstanceIds.getFirst().id()).isEqualTo(c7Instance.getId());
   }
 
   @Test
